move resetAPI call
